requirements.txt
pyyaml
spacy
jinja2
python-docx
weasyprint